code/requirements.txt
bottle
PyYAML
gunicorn